# contracts/openapi.json
{"components":{"schemas":{"ApiError":{"additionalProperties":false,"properties":{"error":{"$ref":"#/components/schemas/ErrorDetail"},"request_id":{"title":"Request Id","type":"string"}},"required":["error","request_id"],"title":"ApiError","type":"object"},"Capabilities":{"additionalProperties":false,"properties":{"global_remaining":{"title":"Global Remaining","type":"integer"},"live_available":{"title":"Live Available","type":"boolean"},"methods":{"items":{"$ref":"#/components/schemas/Method"},"title":"Methods","type":"array"},"next_attempt_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Next Attempt At"},"reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reason"},"visitor_remaining":{"title":"Visitor Remaining","type":"integer"}},"required":["live_available","reason","methods","visitor_remaining","global_remaining","next_attempt_at"],"title":"Capabilities","type":"object"},"Citation":{"additionalProperties":false,"properties":{"citation_id":{"maxLength":160,"pattern":"^[a-z0-9][a-z0-9._-]*$","title":"Citation Id","type":"string"},"quote":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Quote"},"resolvable":{"title":"Resolvable","type":"boolean"},"segment_id":{"title":"Segment Id","type":"string"}},"required":["citation_id","segment_id","resolvable","quote"],"title":"Citation","type":"object"},"ClinicalNote":{"additionalProperties":false,"properties":{"canonical_text":{"title":"Canonical Text","type":"string"},"case_id":{"maxLength":160,"pattern":"^[a-z0-9][a-z0-9._-]*$","title":"Case Id","type":"string"},"note_sha256":{"pattern":"^[0-9a-f]{64}$","title":"Note Sha256","type":"string"},"sections":{"items":{"$ref":"#/components/schemas/NoteSection"},"title":"Sections","type":"array"}},"required":["case_id","sections","canonical_text","note_sha256"],"title":"ClinicalNote","type":"object"},"ConsultationCase":{"additionalProperties":false,"properties":{"case_id":{"maxLength":160,"pattern":"^[a-z0-9][a-z0-9._-]*$","title":"Case Id","type":"string"},"family_id":{"enum":["respiratoire","cardiovasculaire","digestif","urinaire","musculosquelettique","neurologique","dermatologique","suivi-chronique","sante-psychique","prevention"],"title":"Family Id","type":"string"},"focus_tags":{"items":{"type":"string"},"title":"Focus Tags","type":"array"},"group_id":{"maxLength":160,"pattern":"^[a-z0-9][a-z0-9._-]*$","title":"Group Id","type":"string"},"locale":{"const":"fr-FR","title":"Locale","type":"string"},"provenance_id":{"maxLength":160,"pattern":"^[a-z0-9][a-z0-9._-]*$","title":"Provenance Id","type":"string"},"schema_version":{"const":"1.0","title":"Schema Version","type":"string"},"segments":{"items":{"$ref":"#/components/schemas/SourceSegment"},"maxItems":24,"minItems":8,"title":"Segments","type":"array"},"split":{"enum":["dev","test","stress"],"title":"Split","type":"string"},"stress_pair":{"anyOf":[{"$ref":"#/components/schemas/StressPairMetadata"},{"type":"null"}]},"suite":{"enum":["main","stress"],"title":"Suite","type":"string"},"title":{"title":"Title","type":"string"}},"required":["schema_version","case_id","group_id","family_id","suite","split","title","locale","segments","provenance_id","focus_tags","stress_pair"],"title":"ConsultationCase","type":"object"},"ErrorDetail":{"additionalProperties":false,"properties":{"code":{"title":"Code","type":"string"},"message":{"title":"Message","type":"string"},"retry_after_seconds":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Retry After Seconds"}},"required":["code","message","retry_after_seconds"],"title":"ErrorDetail","type":"object"},"ExampleSummary":{"additionalProperties":false,"properties":{"case_id":{"title":"Case Id","type":"string"},"focus":{"title":"Focus","type":"string"},"origins":{"additionalProperties":{"enum":["illustrative","llm_recorded","llm_live"],"type":"string"},"propertyNames":{"$ref":"#/components/schemas/Method"},"title":"Origins","type":"object"},"title":{"title":"Title","type":"string"}},"required":["case_id","title","focus","origins"],"title":"ExampleSummary","type":"object"},"GenerateRequest":{"additionalProperties":false,"properties":{"case_id":{"maxLength":160,"pattern":"^[a-z0-9][a-z0-9._-]*$","title":"Case Id","type":"string"},"method":{"$ref":"#/components/schemas/Method"}},"required":["case_id","method"],"title":"GenerateRequest","type":"object"},"GenerationResult":{"additionalProperties":false,"properties":{"case_id":{"maxLength":160,"pattern":"^[a-z0-9][a-z0-9._-]*$","title":"Case Id","type":"string"},"checks":{"$ref":"#/components/schemas/TechnicalChecks"},"fidelity_metrics":{"anyOf":[{"$ref":"#/components/schemas/NoteEvaluation"},{"type":"null"}]},"metadata":{"$ref":"#/components/schemas/RunMetadata"},"method":{"$ref":"#/components/schemas/Method"},"note":{"$ref":"#/components/schemas/ClinicalNote"},"origin":{"enum":["illustrative","llm_recorded","llm_live"],"title":"Origin","type":"string"},"review":{"$ref":"#/components/schemas/Review"},"run_id":{"maxLength":160,"pattern":"^[a-z0-9][a-z0-9._-]*$","title":"Run Id","type":"string"}},"required":["run_id","case_id","method","origin","note","checks","metadata","review","fidelity_metrics"],"title":"GenerationResult","type":"object"},"Method":{"enum":["direct","structured"],"title":"Method","type":"string"},"MetricValue":{"additionalProperties":false,"properties":{"denominator":{"title":"Denominator","type":"integer"},"numerator":{"title":"Numerator","type":"integer"},"value":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Value"}},"required":["numerator","denominator","value"],"title":"MetricValue","type":"object"},"NoteAssertion":{"additionalProperties":false,"properties":{"assertion_id":{"maxLength":160,"pattern":"^[a-z0-9][a-z0-9._-]*$","title":"Assertion Id","type":"string"},"citations":{"items":{"$ref":"#/components/schemas/Citation"},"maxItems":8,"title":"Citations","type":"array"},"text":{"maxLength":1000,"minLength":1,"title":"Text","type":"string"}},"required":["assertion_id","text","citations"],"title":"NoteAssertion","type":"object"},"NoteEvaluation":{"additionalProperties":false,"properties":{"annotation_sha256":{"pattern":"^[0-9a-f]{64}$","title":"Annotation Sha256","type":"string"},"metrics":{"additionalProperties":{"$ref":"#/components/schemas/MetricValue"},"title":"Metrics","type":"object"},"note_sha256":{"pattern":"^[0-9a-f]{64}$","title":"Note Sha256","type":"string"}},"required":["note_sha256","annotation_sha256","metrics"],"title":"NoteEvaluation","type":"object"},"NoteSection":{"additionalProperties":false,"properties":{"assertions":{"items":{"$ref":"#/components/schemas/NoteAssertion"},"title":"Assertions","type":"array"},"key":{"$ref":"#/components/schemas/SectionKey"},"title":{"title":"Title","type":"string"}},"required":["key","title","assertions"],"title":"NoteSection","type":"object"},"PairedComparison":{"additionalProperties":false,"properties":{"a":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"A"},"b":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"B"},"ci95_high":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Ci95 High"},"ci95_low":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Ci95 Low"},"difference":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Difference"},"n":{"title":"N","type":"integer"},"unit":{"title":"Unit","type":"string"},"warnings":{"items":{"type":"string"},"title":"Warnings","type":"array"}},"required":["a","b","difference","ci95_low","ci95_high","n","unit","warnings"],"title":"PairedComparison","type":"object"},"PublicExample":{"additionalProperties":false,"properties":{"consultation":{"$ref":"#/components/schemas/ConsultationCase"},"results":{"additionalProperties":{"$ref":"#/components/schemas/GenerationResult"},"propertyNames":{"$ref":"#/components/schemas/Method"},"title":"Results","type":"object"}},"required":["consultation","results"],"title":"PublicExample","type":"object"},"PublishedMetrics":{"additionalProperties":false,"properties":{"by_method":{"additionalProperties":{"additionalProperties":{"$ref":"#/components/schemas/MetricValue"},"type":"object"},"propertyNames":{"$ref":"#/components/schemas/Method"},"title":"By Method","type":"object"},"comparisons":{"additionalProperties":{"$ref":"#/components/schemas/PairedComparison"},"title":"Comparisons","type":"object"},"conditional_by_method":{"additionalProperties":{"additionalProperties":{"$ref":"#/components/schemas/MetricValue"},"type":"object"},"propertyNames":{"$ref":"#/components/schemas/Method"},"title":"Conditional By Method","type":"object"},"counts_by_method":{"additionalProperties":{"additionalProperties":{"type":"integer"},"type":"object"},"propertyNames":{"$ref":"#/components/schemas/Method"},"title":"Counts By Method","type":"object"},"failures":{"additionalProperties":{"additionalProperties":{"type":"integer"},"type":"object"},"propertyNames":{"$ref":"#/components/schemas/Method"},"title":"Failures","type":"object"},"stress":{"additionalProperties":{"$ref":"#/components/schemas/MetricValue"},"propertyNames":{"$ref":"#/components/schemas/Method"},"title":"Stress","type":"object"}},"required":["by_method","counts_by_method","conditional_by_method","comparisons","stress","failures"],"title":"PublishedMetrics","type":"object"},"PublishedReport":{"additionalProperties":false,"properties":{"artifact_links":{"additionalProperties":{"type":"string"},"title":"Artifact Links","type":"object"},"cohort_sizes":{"additionalProperties":{"type":"integer"},"title":"Cohort Sizes","type":"object"},"hashes":{"additionalProperties":{"pattern":"^[0-9a-f]{64}$","type":"string"},"title":"Hashes","type":"object"},"limitations":{"items":{"type":"string"},"title":"Limitations","type":"array"},"metrics":{"anyOf":[{"$ref":"#/components/schemas/PublishedMetrics"},{"type":"null"}]},"review_status":{"additionalProperties":{"type":"string"},"title":"Review Status","type":"object"},"schema_version":{"const":"1.0","title":"Schema Version","type":"string"},"status":{"enum":["awaiting_runs","awaiting_annotation","awaiting_reference_review","ai_annotated","human_reviewed"],"title":"Status","type":"string"}},"required":["schema_version","status","cohort_sizes","review_status","metrics","limitations","artifact_links","hashes"],"title":"PublishedReport","type":"object"},"Review":{"additionalProperties":false,"properties":{"artifact_sha256":{"anyOf":[{"pattern":"^[0-9a-f]{64}$","type":"string"},{"type":"null"}],"title":"Artifact Sha256"},"kind":{"enum":["none","ai","human"],"title":"Kind","type":"string"},"reviewed_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reviewed At"},"reviewer_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reviewer Id"}},"required":["kind","reviewer_id","artifact_sha256","reviewed_at"],"title":"Review","type":"object"},"RunMetadata":{"additionalProperties":false,"properties":{"attempts":{"minimum":0.0,"title":"Attempts","type":"integer"},"corpus_sha256":{"pattern":"^[0-9a-f]{64}$","title":"Corpus Sha256","type":"string"},"created_at":{"title":"Created At","type":"string"},"duration_ms":{"anyOf":[{"minimum":0.0,"type":"integer"},{"type":"null"}],"title":"Duration Ms"},"estimated_cost_usd":{"anyOf":[{"minimum":0.0,"type":"number"},{"type":"null"}],"title":"Estimated Cost Usd"},"input_tokens":{"anyOf":[{"minimum":0.0,"type":"integer"},{"type":"null"}],"title":"Input Tokens"},"model_requested":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model Requested"},"model_returned":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model Returned"},"output_tokens":{"anyOf":[{"minimum":0.0,"type":"integer"},{"type":"null"}],"title":"Output Tokens"},"prompt_sha256":{"pattern":"^[0-9a-f]{64}$","title":"Prompt Sha256","type":"string"},"prompt_version":{"title":"Prompt Version","type":"string"},"renderer_sha256":{"pattern":"^[0-9a-f]{64}$","title":"Renderer Sha256","type":"string"},"schema_sha256":{"pattern":"^[0-9a-f]{64}$","title":"Schema Sha256","type":"string"}},"required":["model_requested","model_returned","prompt_version","prompt_sha256","schema_sha256","renderer_sha256","corpus_sha256","created_at","attempts","input_tokens","output_tokens","duration_ms","estimated_cost_usd"],"title":"RunMetadata","type":"object"},"SectionKey":{"enum":["reason","history","background","medications_allergies","observations","assessment","plan"],"title":"SectionKey","type":"string"},"SourceSegment":{"additionalProperties":false,"properties":{"segment_id":{"pattern":"^s[0-9]{3}$","title":"Segment Id","type":"string"},"speaker":{"enum":["patient","clinician","other"],"title":"Speaker","type":"string"},"text":{"maxLength":5000,"minLength":1,"title":"Text","type":"string"}},"required":["segment_id","speaker","text"],"title":"SourceSegment","type":"object"},"StressPairMetadata":{"additionalProperties":false,"properties":{"invariant_fact_keys":{"items":{"maxLength":160,"pattern":"^[a-z0-9][a-z0-9._-]*$","type":"string"},"title":"Invariant Fact Keys","type":"array"},"pair_id":{"maxLength":160,"pattern":"^[a-z0-9][a-z0-9._-]*$","title":"Pair Id","type":"string"},"target_fact_key":{"maxLength":160,"pattern":"^[a-z0-9][a-z0-9._-]*$","title":"Target Fact Key","type":"string"},"variant":{"enum":["positive","negative"],"title":"Variant","type":"string"}},"required":["pair_id","variant","target_fact_key","invariant_fact_keys"],"title":"StressPairMetadata","type":"object"},"TechnicalChecks":{"additionalProperties":false,"properties":{"assertions_without_citation":{"items":{"type":"string"},"title":"Assertions Without Citation","type":"array"},"references_emitted":{"title":"References Emitted","type":"integer"},"references_resolvable":{"title":"References Resolvable","type":"integer"},"schema_valid":{"title":"Schema Valid","type":"boolean"},"unresolved_ids":{"items":{"type":"string"},"title":"Unresolved Ids","type":"array"},"warnings":{"items":{"type":"string"},"title":"Warnings","type":"array"}},"required":["schema_valid","references_emitted","references_resolvable","unresolved_ids","assertions_without_citation","warnings"],"title":"TechnicalChecks","type":"object"}}},"info":{"title":"MediNote","version":"1.0"},"openapi":"3.1.0","paths":{"/api/capabilities":{"get":{"operationId":"get_capabilities_api_capabilities_get","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Capabilities"}}},"description":"Successful Response"},"403":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Forbidden"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Not Found"},"413":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Request Entity Too Large"},"415":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Unsupported Media Type"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Unprocessable Entity"},"429":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Too Many Requests"},"502":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Bad Gateway"},"503":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Service Unavailable"},"504":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Gateway Timeout"}},"summary":"Get Capabilities"}},"/api/examples":{"get":{"operationId":"list_examples_api_examples_get","responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ExampleSummary"},"title":"Response List Examples Api Examples Get","type":"array"}}},"description":"Successful Response"},"403":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Forbidden"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Not Found"},"413":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Request Entity Too Large"},"415":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Unsupported Media Type"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Unprocessable Entity"},"429":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Too Many Requests"},"502":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Bad Gateway"},"503":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Service Unavailable"},"504":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Gateway Timeout"}},"summary":"List Examples"}},"/api/examples/{case_id}":{"get":{"operationId":"get_example_api_examples__case_id__get","parameters":[{"in":"path","name":"case_id","required":true,"schema":{"title":"Case Id","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublicExample"}}},"description":"Successful Response"},"403":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Forbidden"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Not Found"},"413":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Request Entity Too Large"},"415":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Unsupported Media Type"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Unprocessable Entity"},"429":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Too Many Requests"},"502":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Bad Gateway"},"503":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Service Unavailable"},"504":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Gateway Timeout"}},"summary":"Get Example"}},"/api/generate":{"post":{"operationId":"generate_note_api_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateRequest"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerationResult"}}},"description":"Successful Response"},"403":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Forbidden"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Not Found"},"413":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Request Entity Too Large"},"415":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Unsupported Media Type"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Unprocessable Entity"},"429":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Too Many Requests"},"502":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Bad Gateway"},"503":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Service Unavailable"},"504":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Gateway Timeout"}},"summary":"Generate Note"}},"/api/health/live":{"get":{"operationId":"live_api_health_live_get","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"title":"Response Live Api Health Live Get","type":"object"}}},"description":"Successful Response"}},"summary":"Live"}},"/api/health/ready":{"get":{"operationId":"ready_api_health_ready_get","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"title":"Response Ready Api Health Ready Get","type":"object"}}},"description":"Successful Response"},"403":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Forbidden"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Not Found"},"413":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Request Entity Too Large"},"415":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Unsupported Media Type"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Unprocessable Entity"},"429":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Too Many Requests"},"502":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Bad Gateway"},"503":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Service Unavailable"},"504":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Gateway Timeout"}},"summary":"Ready"}},"/api/report":{"get":{"operationId":"get_report_api_report_get","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublishedReport"}}},"description":"Successful Response"},"403":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Forbidden"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Not Found"},"413":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Request Entity Too Large"},"415":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Unsupported Media Type"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Unprocessable Entity"},"429":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Too Many Requests"},"502":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Bad Gateway"},"503":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Service Unavailable"},"504":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}},"description":"Gateway Timeout"}},"summary":"Get Report"}}}}